Moved node
